requirements.txt
pytorch
pyformlang
tqdm
scikit-learn
matplotlib